name of slice refactoring
